Reorg attitude functions into gnc_utils.py
